requirements.txt
transmission-rpc==3.0.0a3
python-telegram-bot==12.8
